requirements.txt
streamlit==1.33.0
langchain==0.1.20
langchain-community==0.0.38
transformers==4.40.0
accelerate==0.29.3
torch==2.1.0
faiss-cpu==1.8.0
pymupdf==1.24.0
python-docx==1.1.0
sentencepiece
protobuf
huggingface_hub
